# Optimize TypeConversionExtensions.ConvertTo