- Change GetTypeConverterTypeName to return the Type directly, so we aren't calling Type.AssemblyQualifiedName, and then Type.GetType.
- Cache TypeConverters from the "toType". This saves allocating toType.CustomAttributes, and allocating a new TypeConverter every time this is called with the same type, e.g. Color.

## Benchmark Results

The easiest public API I could find to exercise this path was Setter.ProvideValue. But this shows up in a startup profile for AppThemeBindingExtension.ProvideValue:

![image](https://user-images.githubusercontent.com/8291187/160728686-666cb6b5-b9f1-4ee1-b4ad-82b9f5fa3d41.png)

### Before

|       Method |     Mean |    Error |   StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------- |---------:|---------:|---------:|-------:|------:|------:|----------:|
| ProvideValue | 10.71 us | 0.189 us | 0.232 us | 0.5035 |     - |     - |   3.11 KB |

### After

|       Method |     Mean |   Error |  StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------- |---------:|--------:|--------:|-------:|------:|------:|----------:|
| ProvideValue | 309.4 ns | 4.65 ns | 4.13 ns | 0.0353 |     - |     - |     224 B |